+++ db/migrate/20210424231757_add_eligible_itt_subject_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddEligibleIttSubjectToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :eligible_itt_subject, :integer
+  end
+end