+++ spec/features/early_career_payments_claim_spec.rb
@@ -47,7 +47,14 @@ RSpec.feature "Teacher Early Career Payments claims" do
 
     expect(claim.eligibility.reload.pgitt_or_ugitt_course).to eq "postgraduate"
 
-    # TODO [PAGE 10] - Which subject did you do your undergraduate ITT in
+    # [PAGE 10] - Which subject did you do your undergraduate ITT in
+    expect(page).to have_text(I18n.t("early_career_payments.questions.eligible_itt_subject", ug_or_pg: claim.eligibility.pgitt_or_ugitt_course))
+
+    choose "Mathematics"
+    click_on "Continue"
+
+    expect(claim.eligibility.reload.eligible_itt_subject).to eql "mathematics"
+
     # TODO [PAGE 11] - Which subject did you do your postgraduate ITT in
     # TODO [PAGE 12] - Do you teach maths now
     # TODO [PAGE 13] - In what academic year did you start your undergraduate ITT
@@ -135,21 +142,5 @@ RSpec.feature "Teacher Early Career Payments claims" do
 
     # [PAGE 08] - Are you currently subject to dsiciplinary action
     expect(page).to have_text(I18n.t("early_career_payments.questions.disciplinary_action"))
-
-    choose "No"
-    click_on "Continue"
-
-    expect(claim.eligibility.reload.subject_to_disciplinary_action).to eql false
-
-    # [PAGE 09] - Did you do a postgraduate ITT course or undergraduate ITT course
-    expect(page).to have_text(I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"))
-
-    choose "Postgraduate"
-    click_on "Continue"
-
-    expect(claim.eligibility.reload.pgitt_or_ugitt_course).to eq "postgraduate"
-
-    # TODO [PAGE 37] - Check your answers before sending your application
-    expect(page).to have_text("Check your answers before sending your application")
   end
 end