+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -5,9 +5,9 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
     {
       nqt_in_academic_year_after_itt: true,
       employed_as_supply_teacher: false,
-      employed_directly: true,
       subject_to_disciplinary_action: false,
-      pgitt_or_ugitt_course: :postgraduate
+      pgitt_or_ugitt_course: :postgraduate,
+      eligible_itt_subject: :chemistry
     }
   end
   let(:eligibility) { claim.eligibility }
@@ -20,7 +20,12 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
       [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
       [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"],
       [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
-      [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"]
+      [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"],
+      [
+        I18n.t("early_career_payments.questions.eligible_itt_subject", ug_or_pg: eligibility.pgitt_or_ugitt_course),
+        "Chemistry",
+        "eligible-itt-subject"
+      ]
     ]
 
     expect(presenter.answers).to eq(expected_answers)
@@ -34,7 +39,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         has_entire_term_contract: true,
         employed_directly: true,
         subject_to_disciplinary_action: false,
-        pgitt_or_ugitt_course: :postgraduate
+        pgitt_or_ugitt_course: :postgraduate,
+        eligible_itt_subject: :modern_foreign_languages
       }
     end
 
@@ -45,7 +51,16 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         [I18n.t("early_career_payments.questions.has_entire_term_contract"), "Yes", "entire-term-contract"],
         [I18n.t("early_career_payments.questions.employed_directly"), "Yes", "employed-directly"],
         [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
-        [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"]
+        [
+          I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"),
+          "Postgraduate",
+          "postgraduate-itt-or-undergraduate-itt-course"
+        ],
+        [
+          I18n.t("early_career_payments.questions.eligible_itt_subject", ug_or_pg: eligibility.pgitt_or_ugitt_course),
+          "Modern foreign languages",
+          "eligible-itt-subject"
+        ]
       ]
 
       expect(presenter.answers).to eq(expected_answers)