+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog]
       or longer?'
     - Add 'Are you currently subject to disciplinary action?'
     - Add 'Did you do a postgraduate ITT course or an undergraduate ITT course?'
+    - Add 'Which subject did you do your undergraduate ITT in?'
     - Add 'You are not eligible for an early careers payment'
 - Automatically update admin claim tasks with DQT API